+++ app/src/main/java/com/example/kotlinweatherapplication/openweathermap/forecast_models/ForecastWeatherResponse.kt
@@ -8,5 +8,5 @@ data class ForecastWeatherResponse(
     val lat: Double,
     val lon: Double,
     val timezone: String,
-    val timezone_offset: Int
+    val timezone_offset: Double
 )
