+++ app/src/main/java/com/example/kotlinweatherapplication/openweathermap/forecast_models/Hourly.kt
@@ -1,18 +1,18 @@
 package com.example.kotlinweatherapplication.openweathermap.forecast_models
 
 data class Hourly(
-    val clouds: Int,
+    val clouds: Double,
     val dew_point: Double,
     val dt: Long,
     val feels_like: Double,
-    val humidity: Int,
+    val humidity: Double,
     val pop: Double,
-    val pressure: Int,
+    val pressure: Double,
     val temp: Double,
     val uvi: Double,
-    val visibility: Int,
+    val visibility: Double,
     val weather: List<WeatherXX>,
-    val wind_deg: Int,
+    val wind_deg: Double,
     val wind_gust: Double,
     val wind_speed: Double
 )
